[MODIFY]: moved the Rectangle3D class files into Math folder
